Libraries Config, Base Controller, DocumentValidator Added

+++ GenericApi.Services/Services/BaseService.cs
@@ -1,4 +1,7 @@
 ﻿using AutoMapper;
+using FluentValidation;
+using GenericApi.Bl.Extensions;
+using GenericApi.Core.Abstract;
 using GenericApi.Core.BaseModel;
 using GenericApi.Model.Repositories;
 using Microsoft.EntityFrameworkCore;
@@ -11,9 +14,9 @@ namespace GenericApi.Services.Services
     {
         Task<IEnumerable<TDto>> GetAllAsync();
         Task<TDto> GetByIdAsync(int id);
-        Task<TDto> AddAsync(TDto dto);
-        Task<TDto> UpdateAsync(int id, TDto dto);
-        Task<TDto> DeleteByIdAsync(int id);
+        Task<IEntityOperationResult<TDto>> AddAsync(TDto dto);
+        Task<IEntityOperationResult<TDto>> UpdateAsync(int id, TDto dto);
+        Task<IEntityOperationResult<TDto>> DeleteByIdAsync(int id);
     }
     public class BaseService<TEntity, TDto> : IBaseService<TEntity, TDto>
         where TEntity : class, IBase
@@ -21,10 +24,12 @@ namespace GenericApi.Services.Services
     {
         protected readonly IMapper _mapper;
         protected readonly IBaseRepository<TEntity> _repository;
-        public BaseService(IBaseRepository<TEntity> repository, IMapper mapper)
+        protected readonly IValidator<TDto> _validator;
+        public BaseService(IBaseRepository<TEntity> repository, IMapper mapper, IValidator<TDto> validator)
         {
             _repository = repository;
             _mapper = mapper;
+            _validator = validator;
         }
         public async Task<IEnumerable<TDto>> GetAllAsync()
         {
@@ -38,30 +43,50 @@ namespace GenericApi.Services.Services
             var dto = _mapper.Map<TDto>(entity);
             return dto;
         }
-        public async Task<TDto> AddAsync(TDto dto)
+        public async Task<IEntityOperationResult<TDto>> AddAsync(TDto dto)
         {
-            TEntity entity = _mapper.Map<TEntity>(dto);
-            var result = await _repository.Add(entity);
-            _mapper.Map(result, dto);
+            var validationResult = _validator.Validate(dto);
+            if (validationResult.IsValid is false) 
+                return validationResult.ToOperationResult<TDto>();
 
-            return dto;
+            TEntity entity = _mapper.Map<TEntity>(dto);
+            var entityResult = await _repository.Add(entity);
+
+            _mapper.Map(entityResult, dto);
+
+            var result = dto.ToOperationResult();
+            return result;
         }
 
-        public async Task<TDto> UpdateAsync(int id, TDto dto)
+        public async Task<IEntityOperationResult<TDto>> UpdateAsync(int id, TDto dto)
         {
+            var validationResult = _validator.Validate(dto);
+            if (validationResult.IsValid is false) 
+                return validationResult.ToOperationResult<TDto>();
+
             var entity = await _repository.Get(id);
+            if (entity is null) return null;
+
             _mapper.Map(dto, entity);
 
-            var result = await _repository.Update(entity);
-            _mapper.Map(result, dto);
-            return dto;
+            entity = await _repository.Update(entity);
+            _mapper.Map(entity, dto);
+
+            var result = dto.ToOperationResult();
+            return result;
         }
 
-        public async Task<TDto> DeleteByIdAsync(int id)
+        public async Task<IEntityOperationResult<TDto>> DeleteByIdAsync(int id)
         {
-            var entity = await _repository.Delete(id);
+            var entity = await _repository.Get(id);
+
+            if (entity is null) return null;
+
+            entity = await _repository.Delete(id);
             var dto = _mapper.Map<TDto>(entity);
-            return dto;
+
+            var result = dto.ToOperationResult();
+            return result;
         }
     }
 }
